#31: Removing previous multitasking implementations
