+++ src/simod/multitasking/__init__.py
@@ -1,0 +1,1 @@
+from .multitasking import adjust_durations
